Introduce unique component for BDF
