+++ Sofa/Component/ODESolver/Backward/CMakeLists.txt
@@ -7,9 +7,7 @@ set(HEADER_FILES
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/config.h.in
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/init.h
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/BaseLinearMultiStepMethod.h
-    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/BDF1.h
-    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/BDF1OdeSolver.h
-    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/BDF2OdeSolver.h
+    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/BDFOdeSolver.h
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/EulerImplicitSolver.h
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewmarkImplicitSolver.h
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewtonRaphsonSolver.h
@@ -29,9 +27,8 @@ set(HEADER_FILES
 
 set(SOURCE_FILES
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/init.cpp
-    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/BDF1.cpp
-    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/BDF1OdeSolver.cpp
-    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/BDF2OdeSolver.cpp
+    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/BaseLinearMultiStepMethod.cpp
+    ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/BDFOdeSolver.cpp
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/EulerImplicitSolver.cpp
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewmarkImplicitSolver.cpp
     ${SOFACOMPONENTODESOLVERBACKWARD_SOURCE_DIR}/NewtonRaphsonSolver.cpp
