+++ Sofa/Component/ODESolver/Backward/src/sofa/component/odesolver/backward/BDFOdeSolver.cpp
@@ -19,26 +19,68 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#pragma once
+#include <sofa/component/odesolver/backward/BDFOdeSolver.h>
 
-#include <sofa/component/odesolver/backward/config.h>
-#include <sofa/component/odesolver/backward/BaseLinearMultiStepMethod.h>
+#include <sofa/core/ObjectFactory.h>
 
 namespace sofa::component::odesolver::backward
 {
 
-struct SOFA_COMPONENT_ODESOLVER_BACKWARD_API BDF1Parameters
+void registerBDFOdeSolver(sofa::core::ObjectFactory* factory)
 {
-    static constexpr std::size_t Order = 1;
-    static constexpr sofa::type::fixed_array<SReal, Order + 1> a_coef {-1, 1};
-    static constexpr sofa::type::fixed_array<SReal, Order + 1> b_coef {0, 1};
-};
-
-class SOFA_COMPONENT_ODESOLVER_BACKWARD_API BDF1OdeSolver :
-public BaseLinearMultiStepMethod<BDF1Parameters>
-{
-public:
-    SOFA_CLASS(BDF1OdeSolver, BaseLinearMultiStepMethod<BDF1Parameters>);
-};
-
+    factory->registerObjects(
+        core::ObjectRegistrationData("Velocity-based ODE solver using Backward Differentiation Formula (BDF), at any order, supporting variable time step size.")
+            .add<BDFOdeSolver>());
 }
+
+void BDFOdeSolver::recomputeCoefficients(std::size_t order, SReal dt)
+{
+    assert(order > 0);
+
+    m_a_coef.resize(order+1);
+    m_b_coef.resize(order+1);
+
+    /**
+     * Computation of the derivative of the Lagrange inteperpolation polynomials
+     */
+    for (std::size_t j = 0; j < m_a_coef.size(); ++j)
+    {
+        auto& a_j = m_a_coef[j];
+
+        a_j = 0;
+        for (std::size_t i = 0; i < order+1; ++i)
+        {
+            if (i != j)
+            {
+                SReal product = 1_sreal;
+                for (std::size_t m = 0; m < order + 1; ++m)
+                {
+                    if (m != i && m != j)
+                    {
+                        product *= (m_timeList[order] - m_timeList[m]) / (m_timeList[j] - m_timeList[m]);
+                    }
+                }
+                a_j += product / (m_timeList[j] - m_timeList[i]);
+            }
+        }
+    }
+
+    for (SReal& j : m_a_coef)
+    {
+        j *= dt;
+    }
+
+    for (std::size_t j = 0; j < m_a_coef.size() - 1; ++j)
+    {
+        m_a_coef[j] /= m_a_coef[order];
+    }
+    m_b_coef[order] = 1 / m_a_coef[order];
+    m_a_coef[order] = 1;
+
+    for (std::size_t j = 0; j < m_b_coef.size() - 1; ++j)
+    {
+        m_b_coef[j] = 0;
+    }
+}
+
+}  // namespace sofa::component::odesolver::backward
