--- examples/Component/ODESolver/Backward/BDF1OdeSolver_spring3d.scn
@@ -1,32 +1,0 @@
-<Node name="root" gravity="0 -9.81 0" dt="0.01">
-    <RequiredPlugin name="Sofa.Component.Constraint.Projective"/> <!-- Needed to use components [FixedProjectiveConstraint] -->
-    <RequiredPlugin name="Sofa.Component.LinearSolver.Direct"/> <!-- Needed to use components [EigenSimplicialLDLT] -->
-    <RequiredPlugin name="Sofa.Component.Mass"/> <!-- Needed to use components [MeshMatrixMass] -->
-    <RequiredPlugin name="Sofa.Component.ODESolver.Backward"/> <!-- Needed to use components [NewtonRaphsonSolver] -->
-    <RequiredPlugin name="Sofa.Component.ODESolver.Integration"/> <!-- Needed to use components [BDF1] -->
-    <RequiredPlugin name="Sofa.Component.SolidMechanics.Spring"/> <!-- Needed to use components [SpringForceField] -->
-    <RequiredPlugin name="Sofa.Component.StateContainer"/> <!-- Needed to use components [MechanicalObject] -->
-    <RequiredPlugin name="Sofa.Component.Visual"/> <!-- Needed to use components [VisualStyle] -->
-    <RequiredPlugin name="Sofa.GL.Component.Rendering3D"/> <!-- Needed to use components [OglCylinderModel] -->
-    <RequiredPlugin name="Sofa.Component.Topology.Container.Dynamic"/> <!-- Needed to use components [EdgeSetTopologyContainer] -->
-
-    <VisualStyle displayFlags="showWireframe showVisual showBehaviorModels showForceFields" />
-    <DefaultAnimationLoop/>
-    <VisualGrid size="2"/>
-    <LineAxis size="2"/>
-
-    <Node name="Newton">
-
-        <BDF1OdeSolver printLog="false" rayleighMass="0.01" rayleighStiffness="0.01"/>
-        <NewtonRaphsonSolver name="newton" printLog="false" maxNbIterationsNewton="4" maxNbIterationsLineSearch="2" absoluteResidualStoppingThreshold="1e-6"/>
-<!--        <EulerImplicitSolver rayleighMass="0.01" rayleighStiffness="0.01"/>-->
-        <EigenSimplicialLDLT template="CompressedRowSparseMatrix"/>
-
-        <MechanicalObject name="dofs" template="Vec3" position="0 0 0  1 1 0 " showObject="true" showObjectScale="20"/>
-        <EdgeSetTopologyContainer name="topologyContainer" edges="0 1" position="@dofs.position"/>
-        <UniformMass totalMass="15" topology="@topologyContainer"/>
-        <SpringForceField spring="0 1 1000 0 1"/>
-        <FixedProjectiveConstraint indices="0" />
-    </Node>
-
-</Node>
