--- examples/Component/ODESolver/Backward/NewtonRaphsonSolver.scn
@@ -1,72 +1,0 @@
-<Node name="root" gravity="-1.8 0 100" dt="0.01">
-
-    <RequiredPlugin name="Sofa.Component.Constraint.Projective"/> <!-- Needed to use components [FixedProjectiveConstraint] -->
-    <RequiredPlugin name="Sofa.Component.Engine.Select"/> <!-- Needed to use components [BoxROI] -->
-    <RequiredPlugin name="Sofa.Component.IO.Mesh"/> <!-- Needed to use components [MeshGmshLoader,MeshOBJLoader] -->
-    <RequiredPlugin name="Sofa.Component.LinearSolver.Direct"/> <!-- Needed to use components [EigenSimplicialLDLT] -->
-    <RequiredPlugin name="Sofa.Component.Mapping.Linear"/> <!-- Needed to use components [BarycentricMapping] -->
-    <RequiredPlugin name="Sofa.Component.Mass"/> <!-- Needed to use components [MeshMatrixMass] -->
-    <RequiredPlugin name="Sofa.Component.ODESolver.Backward"/> <!-- Needed to use components [NewtonRaphsonSolver] -->
-    <RequiredPlugin name="Sofa.Component.ODESolver.Integration"/> <!-- Needed to use components [BDF1] -->
-    <RequiredPlugin name="Sofa.Component.SolidMechanics.FEM.Elastic"/> <!-- Needed to use components [TetrahedronFEMForceField] -->
-    <RequiredPlugin name="Sofa.Component.StateContainer"/> <!-- Needed to use components [MechanicalObject] -->
-    <RequiredPlugin name="Sofa.Component.Topology.Container.Dynamic"/> <!-- Needed to use components [TetrahedronSetGeometryAlgorithms,TetrahedronSetTopologyContainer] -->
-    <RequiredPlugin name="Sofa.Component.Visual"/> <!-- Needed to use components [VisualStyle] -->
-    <RequiredPlugin name="Sofa.GL.Component.Rendering3D"/> <!-- Needed to use components [OglModel] -->
-
-    <VisualStyle displayFlags="showWireframe showVisual showBehaviorModels showForceFields" />
-    <DefaultAnimationLoop/>
-
-    <Node name="DeformableObject">
-
-        <BDF1 printLog="false" rayleighStiffness="0.0" rayleighMass="0.0"/>
-        <NewtonRaphsonSolver name="odeImplicitSolver" printLog="false"
-                             maxNbIterationsNewton="1" absoluteResidualStoppingThreshold="1e-5"
-                             maxNbIterationsLineSearch="5" lineSearchCoefficient="0.5"
-                             absoluteEstimateDifferenceThreshold="1e-5"
-                             relativeEstimateDifferenceThreshold="1e-5"/>
-        <ConstantSparsityPatternSystem template="CompressedRowSparseMatrix" name="A" checkIndices="false"/>
-        <EigenSimplicialLDLT template="CompressedRowSparseMatrix"/>
-
-        <MeshGmshLoader name="loader" filename="mesh/truthcylinder1.msh" />
-        <TetrahedronSetTopologyContainer src="@loader" name="topologyContainer"/>
-        <TetrahedronSetGeometryAlgorithms name="geomAlgo"/>
-        <MechanicalObject src="@loader" />
-        <MeshMatrixMass totalMass="15" topology="@topologyContainer"/>
-
-        <BoxROI name="box" box="-10 -20 -10  10 -17.5 10" drawBoxes="true"/>
-        <FixedProjectiveConstraint indices="@box.indices" />
-        <TetrahedronFEMForceField name="FEM" youngModulus="10000" poissonRatio="0.49" method="large" />
-<!--        <MeshSpringForceField name="springs" tetrasStiffness="10000"/>-->
-
-        <Node name="visual">
-            <MeshOBJLoader name="meshLoader_0" filename="mesh/truthcylinder1.obj" handleSeams="1" />
-            <OglModel name="Visual" src="@meshLoader_0" color="red" />
-            <BarycentricMapping input="@.." output="@Visual" />
-        </Node>
-    </Node>
-
-    <Node name="RefDeformableObject">
-        
-        <EulerImplicitSolver/>
-        <ConstantSparsityPatternSystem template="CompressedRowSparseMatrix" name="A" checkIndices="false"/>
-        <EigenSimplicialLDLT template="CompressedRowSparseMatrix"/>
-
-        <MeshGmshLoader name="loader" filename="mesh/truthcylinder1.msh" />
-        <TetrahedronSetTopologyContainer src="@loader" name="topologyContainer"/>
-        <TetrahedronSetGeometryAlgorithms name="geomAlgo"/>
-        <MechanicalObject src="@loader" />
-        <MeshMatrixMass totalMass="15" topology="@topologyContainer"/>
-
-        <BoxROI name="box" box="-10 -20 -10  10 -17.5 10" drawBoxes="true"/>
-        <FixedProjectiveConstraint indices="@box.indices" />
-        <TetrahedronFEMForceField name="FEM" youngModulus="10000" poissonRatio="0.49" method="large" />
-        <!--        <MeshSpringForceField name="springs" tetrasStiffness="10000"/>-->
-
-        <Node name="visual">
-            <MeshOBJLoader name="meshLoader_0" filename="mesh/truthcylinder1.obj" handleSeams="1" />
-            <OglModel name="Visual" src="@meshLoader_0" color="green" />
-            <BarycentricMapping input="@.." output="@Visual" />
-        </Node>
-    </Node>
-</Node>
