--- examples/Component/ODESolver/Backward/NewtonRaphsonSolver_spring.scn
@@ -1,52 +1,0 @@
-<Node name="root" gravity="0 0 -9.81" dt="0.01">
-    <RequiredPlugin name="Sofa.Component.Constraint.Projective"/> <!-- Needed to use components [FixedProjectiveConstraint] -->
-    <RequiredPlugin name="Sofa.Component.LinearSolver.Direct"/> <!-- Needed to use components [EigenSimplicialLDLT] -->
-    <RequiredPlugin name="Sofa.Component.Mass"/> <!-- Needed to use components [MeshMatrixMass] -->
-    <RequiredPlugin name="Sofa.Component.ODESolver.Backward"/> <!-- Needed to use components [NewtonRaphsonSolver] -->
-    <RequiredPlugin name="Sofa.Component.ODESolver.Integration"/> <!-- Needed to use components [BDF1] -->
-    <RequiredPlugin name="Sofa.Component.SolidMechanics.Spring"/> <!-- Needed to use components [SpringForceField] -->
-    <RequiredPlugin name="Sofa.Component.StateContainer"/> <!-- Needed to use components [MechanicalObject] -->
-    <RequiredPlugin name="Sofa.Component.Visual"/> <!-- Needed to use components [VisualStyle] -->
-    <RequiredPlugin name="Sofa.GL.Component.Rendering3D"/> <!-- Needed to use components [OglCylinderModel] -->
-    <RequiredPlugin name="Sofa.Component.Topology.Container.Dynamic"/> <!-- Needed to use components [EdgeSetTopologyContainer] -->
-
-
-    <VisualStyle displayFlags="showWireframe showVisual showBehaviorModels showForceFields" />
-    <DefaultAnimationLoop/>
-    <VisualGrid size="2"/>
-    <LineAxis size="2"/>
-
-    <Node name="Newton">
-
-        <BDF1 printLog="false" rayleighStiffness="0.0" rayleighMass="0.0"/>
-        <NewtonRaphsonSolver name="odeImplicitSolver" printLog="false" maxNbIterationsNewton="10" 
-                             absoluteResidualStoppingThreshold="1e-8" relativeSuccessiveStoppingThreshold="1e-6"
-                             relativeEstimateDifferenceThreshold="1e-10"
-                             maxNbIterationsLineSearch="10" lineSearchCoefficient="0.5"/>
-        <EigenSimplicialLDLT template="CompressedRowSparseMatrixMat3x3"/>
-
-        <MechanicalObject name="dofs" position="0 0 0  1 1 0" showObject="true"/>
-        <EdgeSetTopologyContainer name="topologyContainer" edges="0 1" position="@dofs.position"/>
-        <UniformMass totalMass="15" topology="@topologyContainer"/>
-        <SpringForceField spring="0 1 1000 0 1"/>
-
-        <FixedProjectiveConstraint indices="0" />
-
-<!--        <OglCylinderModel position="@dofs.position"  edges="0 1" radius="0.1"/>-->
-    </Node>
-
-<!--    <Node name="ForceLinearization">-->
-
-<!--        <EulerImplicitSolver printLog="false" rayleighStiffness="0.01" rayleighMass="0.01"/>-->
-<!--        <EigenSimplicialLDLT template="CompressedRowSparseMatrixMat3x3"/>-->
-
-<!--        <MechanicalObject name="dofs" position="0 0 0  1 1 0"/>-->
-<!--        <EdgeSetTopologyContainer name="topologyContainer" edges="0 1" position="@dofs.position"/>-->
-<!--        <UniformMass totalMass="15" topology="@topologyContainer"/>-->
-<!--        <SpringForceField spring="0 1 1000 0 1"/>-->
-
-<!--        <FixedProjectiveConstraint indices="0" />-->
-
-<!--        <OglCylinderModel position="@dofs.position"  edges="0 1" radius="0.1"/>-->
-<!--    </Node>-->
-</Node>
