--- examples/Component/ODESolver/Backward/NewtonRaphsonSolver_spring1d.scn
@@ -1,31 +1,0 @@
-<Node name="root" gravity="-1 0 0" dt="0.01">
-    <RequiredPlugin name="Sofa.Component.Constraint.Projective"/> <!-- Needed to use components [FixedProjectiveConstraint] -->
-    <RequiredPlugin name="Sofa.Component.LinearSolver.Direct"/> <!-- Needed to use components [EigenSimplicialLDLT] -->
-    <RequiredPlugin name="Sofa.Component.Mass"/> <!-- Needed to use components [MeshMatrixMass] -->
-    <RequiredPlugin name="Sofa.Component.ODESolver.Backward"/> <!-- Needed to use components [NewtonRaphsonSolver] -->
-    <RequiredPlugin name="Sofa.Component.ODESolver.Integration"/> <!-- Needed to use components [BDF1] -->
-    <RequiredPlugin name="Sofa.Component.SolidMechanics.Spring"/> <!-- Needed to use components [SpringForceField] -->
-    <RequiredPlugin name="Sofa.Component.StateContainer"/> <!-- Needed to use components [MechanicalObject] -->
-    <RequiredPlugin name="Sofa.Component.Visual"/> <!-- Needed to use components [VisualStyle] -->
-    <RequiredPlugin name="Sofa.GL.Component.Rendering3D"/> <!-- Needed to use components [OglCylinderModel] -->
-    <RequiredPlugin name="Sofa.Component.Topology.Container.Dynamic"/> <!-- Needed to use components [EdgeSetTopologyContainer] -->
-
-
-    <VisualStyle displayFlags="showWireframe showVisual showBehaviorModels showForceFields" />
-    <DefaultAnimationLoop/>
-    <VisualGrid size="2"/>
-    <LineAxis size="2"/>
-
-    <Node name="Newton">
-
-        <BDF1 printLog="true"/>
-        <NewtonRaphsonSolver name="odeImplicitSolver" printLog="true" maxNbIterationsNewton="4" maxNbIterationsLineSearch="2" absoluteResidualStoppingThreshold="1e-6"/>
-        <EigenSimplicialLDLT template="CompressedRowSparseMatrix"/>
-
-        <MechanicalObject name="dofs" template="Vec1" position="1 " rest_position="0" showObject="true"/>
-        <PointSetTopologyContainer name="topologyContainer" position="1"/>
-        <UniformMass totalMass="1" topology="@topologyContainer"/>
-        <RestShapeSpringsForceField points="0"/>
-    </Node>
-
-</Node>
